test(lint): move directive integration coverage
